音声入力処理を分離し、連続入力を受け付けるよう対応。
